Test socket non-blocking

+++ skipi_receiver.py
@@ -30,13 +30,16 @@ with open(PID_FILE, 'r') as f:
 print "Receiver - Have pid: ", str(os.getpid())
 print "Receiver - Got pid: ", skipi_pid
 
+sock.setblocking(0)
+
 while os.path.isfile(PID_FILE):
     data, addr = sock.recvfrom(MSG_MAX_LEN) # buffer size is 20 bytes
-    with open(LED_MODE_FILE, 'w',0) as fd:
-        fd.write(data)
-        fd.write('\n')
-        fd.flush()
-        fd.close()
-    sleep(5)
-    os.kill(skipi_pid, signal.SIGUSR1)
-    print "Receiver - Got data: ", data
+    if data:
+        with open(LED_MODE_FILE, 'w',0) as fd:
+            fd.write(data)
+            fd.write('\n')
+            fd.flush()
+            fd.close()
+            sleep(5)
+            os.kill(skipi_pid, signal.SIGUSR1)
+            print "Receiver - Got data: ", data
